[OGUI-367] Extract checkers in separate component
